moved document to utils
